Move dj gateway link to gateway module
